+++ lib/cogynt_workstation_ingest/utils/tasks/delete_drilldown_data_task.ex
@@ -49,15 +49,22 @@ defmodule CogyntWorkstationIngest.Utils.Tasks.DeleteDrilldownDataTask do
         )
 
         # Delete topics for worker
-        delete_topic_result =
-          KafkaEx.delete_topics([Config.topic_sols(), Config.topic_sol_events()],
-            worker_name: worker_name
-          )
+        if Process.whereis(worker_name) != nil do
+          delete_topic_result =
+            KafkaEx.delete_topics([Config.topic_sols(), Config.topic_sol_events()],
+              worker_name: worker_name
+            )
 
-        CogyntLogger.info(
-          "#{__MODULE__}",
-          "Deleted Drilldown Topics result: #{inspect(delete_topic_result, pretty: true)}"
-        )
+          CogyntLogger.info(
+            "#{__MODULE__}",
+            "Deleted Drilldown Topics result: #{inspect(delete_topic_result, pretty: true)}"
+          )
+        else
+          CogyntLogger.info(
+            "#{__MODULE__}",
+            "Deleted Drilldown Topics result: No PID associated to #{worker_name}"
+          )
+        end
       end
 
       CogyntLogger.info("#{__MODULE__}", "Starting resetting of drilldown data")
@@ -102,8 +109,15 @@ defmodule CogyntWorkstationIngest.Utils.Tasks.DeleteDrilldownDataTask do
 
   defp reset_cached_data() do
     Redis.key_delete("dcgid")
-    {:ok, keys} = Redis.keys_by_pattern("dmi:*")
-    Redis.key_delete_pipeline(keys)
+
+    case Redis.keys_by_pattern("dmi:*") do
+      {:ok, []} ->
+        nil
+
+      {:ok, keys} ->
+        Redis.key_delete_pipeline(keys)
+    end
+
     DrilldownContext.hard_delete_template_solutions_data()
   end
 
@@ -117,8 +131,8 @@ defmodule CogyntWorkstationIngest.Utils.Tasks.DeleteDrilldownDataTask do
           "Drilldown-#{hashed_brokers}" <> "-" <> consumer_group_id
       end
 
-    if consumer_group_id == false do
-      {:ok, false}
+    if consumer_group_id == nil do
+      {:ok, true}
     else
       case Redis.key_exists?("dmi:#{consumer_group_id}") do
         {:ok, false} ->
